+++ noughtsAndCrosses/input/i_input_manager.h
@@ -8,7 +8,7 @@ namespace ticTacToe
 		virtual ~IInputManager() noexcept {};
 
 		virtual int getKeyPreseed() const = 0;
-		virtual int_fast8_t getDigitPressed() const = 0;
+		virtual int getDigitPressed() const = 0;
 		virtual Side selectSide() const = 0;
 		virtual bool confirmAction() const = 0;
 	};
